chore: moved ui folder to src folder and update paths
