+++ __tests__/faq.test.js
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 import "@testing-library/jest-dom";
-import Faq from "../ui/faq";
+import Faq from "../src/ui/faq";
 
 describe("Faq Component", () => {
   test("renders all AccordionTab headers", () => {
